a=int(input()) b=3.14*a*a c=float("{:.2f}".format(b)) print(c)

+++ Area_of_a_circle.py
@@ -1,5 +1,4 @@
-
-a=float(input())
+a=int(input())
 b=3.14*a*a
 c=float("{:.2f}".format(b))
 print(c)
